// src/api/model/search1Experience.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type Search1Experience =
  (typeof Search1Experience)[keyof typeof Search1Experience];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Search1Experience = {
  NO_EXPERIENCE: "NO_EXPERIENCE",
  BETWEEN_1_AND_3_YEARS: "BETWEEN_1_AND_3_YEARS",
  BETWEEN_3_AND_6_YEARS: "BETWEEN_3_AND_6_YEARS",
  MORE_THAN_6_YEARS: "MORE_THAN_6_YEARS",
} as const;