// src/api/model/search1Schedule.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type Search1Schedule =
  (typeof Search1Schedule)[keyof typeof Search1Schedule];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Search1Schedule = {
  FULL_DAY: "FULL_DAY",
  SHIFT: "SHIFT",
  FLEXIBLE: "FLEXIBLE",
  REMOTE: "REMOTE",
  FLY_IN_FLY_OUT: "FLY_IN_FLY_OUT",
  PARTIAL_DAY: "PARTIAL_DAY",
} as const;