// src/api/model/search1Education.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type Search1Education =
  (typeof Search1Education)[keyof typeof Search1Education];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Search1Education = {
  SECONDARY: "SECONDARY",
  SPECIAL_SECONDARY: "SPECIAL_SECONDARY",
  UNFINISHED_HIGHER: "UNFINISHED_HIGHER",
  HIGHER: "HIGHER",
  BACHELOR: "BACHELOR",
  MASTER: "MASTER",
} as const;